updated ansi-mac.sh - adjusted wget download

+++ ansi-mac.sh
@@ -13,10 +13,12 @@ You are running the OpenShift CDK Lab Installer for MacOS. The following actions
 2. Create ~/git directory and check out cdk-lab into ~/git/cdk-labs
 3. Create ~/bin and extend your PATH to include ~/bin
 4. Install Homebrew
-5. Install wget
-6. Get the latest CDK ( currently nightly builds for cdk-3.1 ) and put it in ~/bin
-7. Install docker-machine-driver-xhyve
-8. Install olab command in ~/bin
+5. Install Homebrew Cask
+6. Install Google Chrome
+7. Install wget
+8. Get the latest CDK ( currently nightly builds for cdk-3.1 ) and put it in ~/bin
+9. Install docker-machine-driver-xhyve
+10. Install olab command in ~/bin
 
 You will need to enter your password for prviledged actions.
 
@@ -65,8 +67,16 @@ test -d ~/bin/cdkshift && echo "~/bin/cdkshift was there already" || mkdir -p ~/
 echo -e "\n${bold}4. Checking / Installing homebrew"
 brew --version &>/dev/null && echo "homebrew already installed" || /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 
+# Install homebrew cask
+echo -e "\n${bold}5. Checking / Installing Homebrew Cask"
+brew list brew-cask &>/dev/null && echo "brew-cask already installed"  || brew install brew-cask
+
+# Install google-chrome
+echo -e "\n${bold}6. Installing Google Chrome"
+brew cask install google-chrome
+
 # Installing wget
-echo -e "\n{$bold}5. Installing wget"
+echo -e "\n{$bold}7. Installing wget"
 brew list wget &>/dev/null && echo "wget already installed" || brew install wget
 
 # Get CDK (ToDo official CDK when released) 
@@ -74,18 +84,18 @@ brew list wget &>/dev/null && echo "wget already installed" || brew install wget
 # - only transfer if newer
 # - link to cdk to preserve existing minishift
 #
-echo -e "\n${bold}6. Getting latest CDK - this can be a slow download of ~400MB"
+echo -e "\n${bold}8. Getting latest CDK - this can be a slow download of ~400MB"
 wget -r --tries=15 --continue -nH --cut-dirs=1 -P ~/bin/cdkshift http://sademo.de/mac/minishift
 test -l ~/bin/cdk || ln -s ~/bin/cdkshift/minishift ~/bin/cdk
 chmod +x ~/bin/cdk
 
 # Install docker-machine-driver-xhyve
-echo -e "\n${bold}7. Installing docker-machine-driver-xhyve"
+echo -e "\n${bold}9. Installing docker-machine-driver-xhyve"
 brew list docker-machine-driver-xhyve &>/dev/null && echo "xhyve was installed" || brew install docker-machine-driver-xhyve
 
 # Install the olab Command in ~/bin
-echo -e "\n${bold}8. Installing olab Script"
-cp ~/git/cdk-labs/olab ~/bin
+echo -e "\n${bold}10. Installing olab Script"
+test -f ~/bin/olab && echo olab already there skipping copy || cp ~/git/cdk-labs/olab ~/bin
 
 # You are ready to roll now
 cat <<ENDMESSAGE
